uint64 -> int

+++ app/services/accounting/api/v1/accountingapi/consumerapi/request.go
@@ -1,13 +1,13 @@
 package consumerapi
 
 type NewConsumerAccountRequest struct {
-	ConsumerID uint64 `json:"consumer_id" yaml:"consumerID"`
+	ConsumerID int `json:"consumer_id" yaml:"consumerID"`
 }
 
 type AddBalanceConsumerAccountRequest struct {
-	Amount int64 `json:"amount" yaml:"amount"`
+	Amount int `json:"amount" yaml:"amount"`
 }
 
 type SubBalanceConsumerAccountRequest struct {
-	Amount int64 `json:"amount" yaml:"amount"`
+	Amount int `json:"amount" yaml:"amount"`
 }
