+++ app/services/accounting/api/v1/accountingapi/consumerapi/response.go
@@ -1,8 +1,8 @@
 package consumerapi
 
 type ConsumerAccountResponse struct {
-	ConsumerID uint64 `json:"consumer_id,omitempty" yaml:"consumer_id,omitempty"`
-	Balance    int64  `json:"balance" yaml:"balance"`
+	ConsumerID int `json:"consumer_id,omitempty" yaml:"consumer_id,omitempty"`
+	Balance    int `json:"balance" yaml:"balance"`
 }
 
 type ReturnAccountingResponseList struct {
