+++ app/services/accounting/api/v1/accountingapi/courierrapi/request.go
@@ -1,13 +1,13 @@
 package courierapi
 
 type NewCourierAccountRequest struct {
-	CourierID uint64 `json:"courier_id" yaml:"courierID"`
+	CourierID int `json:"courier_id" yaml:"courierID"`
 }
 
 type AddBalanceCourierAccountRequest struct {
-	Amount int64 `json:"amount" yaml:"amount"`
+	Amount int `json:"amount" yaml:"amount"`
 }
 
 type SubBalanceCourierAccountRequest struct {
-	Amount int64 `json:"amount" yaml:"amount"`
+	Amount int `json:"amount" yaml:"amount"`
 }
