+++ app/services/accounting/pkg/domain/courieraccount.go
@@ -3,8 +3,8 @@ package domain
 import "time"
 
 type CourierAccount struct {
-	CourierID uint64
-	Balance    int64
-	CreatedAt  time.Time
-	UpdatedAt  time.Time
+	CourierID int
+	Balance   int
+	CreatedAt time.Time
+	UpdatedAt time.Time
 }
