+++ app/services/accounting/pkg/domain/restaurantaccount.go
@@ -3,8 +3,8 @@ package domain
 import "time"
 
 type RestaurantAccount struct {
-	RestaurantID uint64
-	Balance      int64
+	RestaurantID int
+	Balance      int
 	CreatedAt    time.Time
 	UpdatedAt    time.Time
 }
