+++ app/services/accounting/pkg/service/consumerservice/consumerstorage.go
@@ -5,8 +5,8 @@ import "accounting/pkg/domain"
 // ConsumerStorage is the interface for the accounting storage.
 type ConsumerStorage interface {
 	InsertNewConsumerAccount(account domain.ConsumerAccount) (*domain.ConsumerAccount, error)
-	GetConsumerAccountByID(id uint64) (*domain.ConsumerAccount, error)
-	DeleteConsumerAccount(consumerID uint64) error
+	GetConsumerAccountByID(id int) (*domain.ConsumerAccount, error)
+	DeleteConsumerAccount(consumerID int) error
 
 	AddToBalanceConsumerAccount(account domain.ConsumerAccount) (*domain.ConsumerAccount, error)
 	SubFromBalanceConsumerAccount(account domain.ConsumerAccount) (*domain.ConsumerAccount, error)
